requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.6.0
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
requests==2.28.1
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.13
